+++ sample/src/jsMain/kotlin/JSPlugin.kt
@@ -7,22 +7,18 @@ import dev.inmo.navigation.sample.ui.NavigationView
 import dev.inmo.navigation.sample.ui.NavigationViewConfig
 import dev.inmo.micro_utils.koin.singleWithRandomQualifier
 import dev.inmo.micro_utils.startup.plugin.StartPlugin
-import dev.inmo.navigation.core.CookiesNavigationConfigsRepo
 import dev.inmo.navigation.core.NavigationChain
-import dev.inmo.navigation.core.NavigationNode
 import dev.inmo.navigation.core.NavigationNodeFactory
 import dev.inmo.navigation.core.configs.NavigationNodeDefaultConfig
 import dev.inmo.navigation.core.initNavigation
 import dev.inmo.navigation.core.repo.ConfigHolder
 import dev.inmo.navigation.core.repo.NavigationConfigsRepo
 import dev.inmo.navigation.core.urls.OneParameterUrlNavigationConfigsRepo
-import kotlinx.browser.window
 import kotlinx.serialization.PolymorphicSerializer
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.modules.SerializersModule
 import org.koin.core.Koin
 import org.koin.core.module.Module
-import org.w3c.dom.url.URLSearchParams
 
 object JSPlugin : StartPlugin {
     private val Koin.nodeFactory: NavigationNodeFactory<NavigationNodeDefaultConfig>